Remove now-empty PreloaderAssociation override
